+++ src/main/kotlin/com/tegonal/minimalist/generators/arbFromArray.kt
