+++ src/main/kotlin/com/tegonal/minimalist/generators/arbFromEnum.kt
