+++ src/main/kotlin/com/tegonal/minimalist/generators/arbFromList.kt
