+++ src/main/kotlin/com/tegonal/minimalist/generators/arbFromProgression.kt
