+++ src/main/kotlin/com/tegonal/minimalist/generators/orderedTo.kt
@@ -1,5 +1,20 @@
 package com.tegonal.minimalist.generators
 
+import com.tegonal.minimalist.generators.impl.OrderedToArbArgsGenerator
+
+/**
+ * Turns `this` [OrderedArgsGenerator] into a [ArbArgsGenerator].
+ *
+ * @param T The type of values generated by `this` [OrderedArgsGenerator].
+ *
+ * @return The resulting [ArbArgsGenerator].
+ *
+ * @since 2.0.0
+ */
+fun <T> OrderedArgsGenerator<T>.toRandomArgsGenerator(): ArbArgsGenerator<T> =
+	OrderedToArbArgsGenerator(this)
+
+
 /**
  * Generates [size][OrderedArgsGenerator.size] values and puts them into a [List].
  *
@@ -19,3 +34,4 @@ fun <T> OrderedArgsGenerator<T>.toList(): List<T> =
  */
 fun <T> OrderedArgsGenerator<T>.toSet(): Set<T> =
 	generate().take(size).toSet()
+
