--- src/main/kotlin/com/tegonal/minimalist/generators/orderedToRandomArgsGenerator.kt
@@ -1,16 +1,0 @@
-package com.tegonal.minimalist.generators
-
-import com.tegonal.minimalist.generators.impl.OrderedToArbArgsGenerator
-
-/**
- * Turns `this` [OrderedArgsGenerator] into a [ArbArgsGenerator].
- *
- * @param T The type of values generated by `this` [OrderedArgsGenerator].
- *
- * @return The resulting [ArbArgsGenerator].
- *
- * @since 2.0.0
- */
-fun <T> OrderedArgsGenerator<T>.toRandomArgsGenerator(): ArbArgsGenerator<T> =
-	OrderedToArbArgsGenerator(this)
-
